+++ src/Settings.php
@@ -73,7 +73,7 @@ class Settings {
 
 		esc_html_e( 'Rollback to latest release or any outdated, secure release version of WordPress Core.', 'core-rollback' );
 
-		if ( method_exists( 'WP_Theme', 'is_block_theme' ) && wp_get_theme()->is_block_theme() ) {
+		if ( function_exists( 'wp_is_block_theme' ) && wp_is_block_theme() ) {
 			echo '<div class="notice notice-info fade is-dismissible">';
 			echo '<p>' . wp_kses_post( __( '<strong>INFO:</strong> Your site is currently using a block theme. Block themes require at least WordPress 5.9.', 'core-rollback' ) ) . '</p>';
 			echo '</div>';
@@ -136,7 +136,7 @@ class Settings {
 		$items = array_filter(
 			$items,
 			function( $item ) {
-				if ( method_exists( 'WP_Theme', 'is_block_theme' ) && wp_get_theme()->is_block_theme() ) {
+				if ( function_exists( 'wp_is_block_theme' ) && wp_is_block_theme() ) {
 					return version_compare( $item, '5.9', '>=' );
 				}
 				if ( version_compare( phpversion(), '5.7', '<=' ) ) {
@@ -166,7 +166,7 @@ class Settings {
 	private function filter_if_block_themes() {
 		$core  = new Core();
 		$items = array_keys( $core::$core_versions );
-		if ( method_exists( 'WP_Theme', 'is_block_theme' ) && wp_get_theme()->is_block_theme() ) {
+		if ( function_exists( 'wp_is_block_theme' ) && wp_is_block_theme() ) {
 			$items = array_filter(
 				$items,
 				function( $item ) {